requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
billiard==4.1.0
captcha==0.4
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
confusable-homoglyphs==3.2.0
crispy-bootstrap5==0.7
crispy-templates==0.2.0
cryptography==38.0.4
defusedxml==0.7.1
distro==1.8.0
Django==4.1.5
django-allauth==0.52.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-registration==3.3
django-timezone-field==5.0
django-tinymce==3.5.0
django-webpush==0.3.4
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flower==1.2.0
fontawesomefree==6.2.1
gevent==22.10.2
greenlet==2.0.1
http-ece==1.1.0
humanize==4.4.0
idna==3.4
jsonschema==4.17.3
kombu==5.2.4
newsapi-python==0.2.6
oauthlib==3.2.2
packaging==22.0
paramiko==2.12.0
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
py-vapid==1.9.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
pywebpush==1.14.0
pywin32==305
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
tornado==6.2
tzdata==2022.7
urllib3==1.26.13
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.2
zope.event==4.6
zope.interface==5.5.2

APScheduler~=3.9.1.post1